requirements.txt
MySQLdb
cffi==1.9.0
Django==1.10.6
docopt==0.6.2
easywebdav==1.2.0
gevent==1.0.dev0
gevent-websocket==0.9.5
greenlet==0.4.10
jwt==0.3.2
libsass==0.12.3
pkg-resources==0.0.0
pycrypto==2.6.1
PyMySQL==0.6.6
pypycore==0.1
pyserial==3.3
pytz==2016.10
readline==6.2.4.1
requests==2.7.0
six==1.10.0
stackless-python==10.0
suds==0.4
ws4py==0.4.2
